Improvement of the migration in a continuous integration context.
Indeed, before this patch, the implementation was not complete.

+++ PyFunceble/cli/migrators/json2csv/inactive.py
@@ -92,10 +92,20 @@ class InactiveJSON2CSVMigrator(JSON2CSVMigratorBase):
                 "destination": None,
                 "source": None,
                 "tested_at": None,
+                "session_id": None,
             }
+
+            delete_file = True
 
             with file_helper.open("r", encoding="utf-8") as file_stream:
                 for line in file_stream:
+                    if (
+                        self.continuous_integration
+                        and self.continuous_integration.is_time_exceeded()
+                    ):
+                        delete_file = False
+                        break
+
                     line = (
                         line.strip()
                         .replace('"', "")
@@ -139,5 +149,6 @@ class InactiveJSON2CSVMigrator(JSON2CSVMigratorBase):
                         "Added %r into %r", dataset["idna_subject"], self.dataset
                     )
 
-            file_helper.delete()
+            if delete_file:
+                file_helper.delete()
         return self
